Added layout functionality and enabled master HID
New layout system is defined. The layout can be established on the
'config.json'. Two layouts may be set. The 'layout' one maps every
physical key from the keyboard matrix to the desired character.
The 'pc_layout' one maps each character to the keycode the PC understands.
For example, if you set in your operative system the layout 'qwerty-de',
then you should do the same on this parameter. This makes it easier for
the user, because he/she has not to change the layout of his/her computer.