Added macro key processing
A macro key is an special key that does not send keys to the host directly.
They are commonly used to adjust the keyboard. Two macro keys where implemented,
named as LY0 and LY1. These keys allows the user to switch the layer they
are working on, so they can type symbols in an easier way.

We strongly suggest to define these keys in the same position along all the
defined layers, so there is no problem of consistency when pressing and
releasing.